Prevent concurent state emit in iterable player when setting playback
When setting the playback speed for the player the player needs to emit state so downstream components can update with the new playback speed. The iterable player emits the state only when idle, however even when idle the player may be loading blocks and emitting state which would cause concurrent state emits which is invalid in our pipeline.

This change fixes this behavior by tracking whether a state emit is in-flight within the emitState method and waiting for any in-flight state to finish before emitting again.

Fixes: #3375